changed so that all parameters in MathCalculation divide are double

+++ src/main/java/com/example/springdocker/MathCalculation.java
@@ -36,7 +36,7 @@ public class MathCalculation {
      * @param y integer
      * @return x/y float
      */
-    public double divide(int x, int y) {
+    public double divide(double x, double y) {
         return x / y;
 
     }
